+++ dataset/__init__.py
@@ -6,7 +6,7 @@ from .batch import Batch, ArrayBatch, DataFrameBatch
 from .dataset import Dataset
 from .jointdataset import JointDataset, FullDataset
 from .dsindex import DatasetIndex, FilesIndex
-from .preprocess import Preprocessing, action
+from .pipeline import action
 
 if sys.version_info < (3, 4):
     raise ImportError("Dataset module requires Python 3.4 or higher")
